+++ e2e/e2etests/test_spl_withdraw.go
@@ -9,6 +9,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/zeta-chain/node/e2e/runner"
+	"github.com/zeta-chain/node/e2e/utils"
+	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 )
 
 func TestSPLWithdraw(r *runner.E2ERunner, args []string) {
@@ -29,21 +31,25 @@ func TestSPLWithdraw(r *runner.E2ERunner, args []string) {
 		r,
 		-1,
 		withdrawAmount.Cmp(approvedAmount),
-		"Withdrawal amount must be less than the approved amount (1e9)",
+		"Withdrawal amount must be less than the %v",
+		approvedAmount,
 	)
 
 	// load deployer private key
-	privkey, err := solana.PrivateKeyFromBase58(r.Account.SolanaPrivateKey.String())
-	require.NoError(r, err)
+	privkey := r.GetSolanaPrivKey()
 
 	// get receiver ata balance before withdraw
-	receiverAta := r.FindOrCreateAssociatedTokenAccount(privkey, privkey.PublicKey(), r.SPLAddr)
+	receiverAta := r.FindOrCreateAta(privkey, privkey.PublicKey(), r.SPLAddr)
 	receiverBalanceBefore, err := r.SolanaClient.GetTokenAccountBalance(r.Ctx, receiverAta, rpc.CommitmentConfirmed)
 	require.NoError(r, err)
 	r.Logger.Info("receiver balance of SPL before withdraw: %s", receiverBalanceBefore.Value.Amount)
 
 	// withdraw
-	r.WithdrawSPLZRC20(privkey.PublicKey(), withdrawAmount, approvedAmount)
+	tx := r.WithdrawSPLZRC20(privkey.PublicKey(), withdrawAmount, approvedAmount)
+
+	// wait for the cctx to be mined
+	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_OutboundMined)
 
 	// get SPL ZRC20 balance after withdraw
 	zrc20BalanceAfter, err := r.SPLZRC20.BalanceOf(&bind.CallOpts{}, r.EVMAddress())
@@ -56,14 +62,8 @@ func TestSPLWithdraw(r *runner.E2ERunner, args []string) {
 	r.Logger.Info("receiver balance of SPL after withdraw: %s", receiverBalanceAfter.Value.Amount)
 
 	// verify amount is added to receiver ata
-	require.Zero(
-		r,
-		new(
-			big.Int,
-		).Add(withdrawAmount, parseBigInt(r, receiverBalanceBefore.Value.Amount)).
-			Cmp(parseBigInt(r, receiverBalanceAfter.Value.Amount)),
-	)
+	require.EqualValues(r, new(big.Int).Add(withdrawAmount, parseBigInt(r, receiverBalanceBefore.Value.Amount)).String(), parseBigInt(r, receiverBalanceAfter.Value.Amount).String())
 
 	// verify amount is subtracted on zrc20
-	require.Zero(r, new(big.Int).Sub(zrc20BalanceBefore, withdrawAmount).Cmp(zrc20BalanceAfter))
+	require.EqualValues(r, new(big.Int).Sub(zrc20BalanceBefore, withdrawAmount).String(), zrc20BalanceAfter.String())
 }